Add boot animation frames
We're adding the frames so that unofficial builders can use the script and so
make custom fitted boot animations instead of having to use a one size fits all
type of boot animation.

To use the script again is easy!

- Just fire up the build script by typing in '.build/envsetup.sh'
- Once that's fired up type in 'bootanimation X Y' without the '' and replacing X and Y
with your width and height of your boot animation.

So for example if I wanted a boot animation that was 1920x1200, I would type in

bootanimation 1920 1200

After you do that, you let it run and follow the instructions after that. Once you have
your custom fitted boot animation, make sure you grab it and store it if you're building
unofficially. If you're a maintainer, obviously put it in vendor/du and submit a proper
commit with the filter and all in place.

Change-Id: I68aa090b93a46d118e1151d89f31723149203e15